// src/models/weapon-category.ts
export enum WeaponCategory {
  AutoRifle = 'Fusils autommatiques',
  Revolver = 'Revolvers',
  PulseRifle = 'Fusils à impulsion',
  ScoutRifle = 'Fusils d\'éclaireur',
  FusionRifle = 'Fusils à fusion',
  SniperRifle = 'Fusils de précision',
  Shotgun = 'Fusils à pompe',
  MachineGun = 'Mitrailleuses',
  RocketLauncher = 'Lance-roquettes',
  Sidearm = 'Pistolets',
  Sword = 'Épées',
  BeamRifle = 'Fusils à rayon',
  GrenadeLauncher = 'Lance-grenades',
  SubMachineGun = 'Pistolets-mitrailleurs',
  LinearFusionRifle = 'Fusils à fusion linéaire',
  Bow = 'Arcs'
}